requirements.txt
PyQt5
qtawesome
requests
pandas
openpyxl
fpdf2
folium
matplotlib
PyInstaller
